Update editor create call
