+++ src/browser/util.ts
@@ -1,6 +1,7 @@
 import MarkdownIt, { Token } from 'markdown-it';
-import { ContextBridgeAPI } from './interfaces/Bridge';
 import Renderer from 'markdown-it/lib/renderer.mjs';
+import Swal, { SweetAlertIcon } from 'sweetalert2';
+import { ContextBridgeAPI } from './interfaces/Bridge';
 
 /**
  * Generate a random number between two numbers.
@@ -93,3 +94,38 @@ export function selfRender(
 ) {
   return self.renderToken(tokens, idx, options);
 }
+
+/**
+ * Configure a sweetalert2 mixin for toast notifications.
+ */
+const toast: ReturnType<typeof Swal.mixin> = Swal.mixin({
+  toast: true,
+  position: 'bottom-end',
+  showConfirmButton: false,
+  showCloseButton: true,
+  timer: 7500,
+  timerProgressBar: true,
+  showClass: {
+    popup: '',
+  },
+  hideClass: {
+    popup: '',
+  },
+  didOpen: (toast) => {
+    toast.addEventListener('mouseenter', Swal.stopTimer);
+    toast.addEventListener('mouseleave', Swal.resumeTimer);
+  },
+});
+
+export const notify = {
+  /**
+   * Send a toast notification.
+   *
+   * @param icon - the icon for the notification
+   * @param html - the content of the notification
+   */
+  async send(icon: string, html: string) {
+    const title = icon.charAt(0).toUpperCase() + icon.slice(1);
+    await toast.fire({ html, title, icon: icon as SweetAlertIcon });
+  },
+};
